builder enhancements

+++ backend/Jimx.WebAggregator.Persistent.MongoDB/Operations/UpsertMongoUnitOfWork.cs
@@ -1,52 +1,64 @@
-﻿using MongoDB.Driver;
+﻿using Jimx.WebAggregator.Domain.MongoDB;
+using Microsoft.Extensions.Logging;
+using MongoDB.Driver;
 
 namespace Jimx.WebAggregator.Persistent.MongoDB.Operations
 {
 	public class UpsertMongoUnitOfWork<TCollectionItem, TIdentity> : MongoUnitOfWork<TCollectionItem>
 		where TCollectionItem : IMongoEntity
 	{
-		private readonly TCollectionItem[] _updatedCollection;
+		private readonly IEnumerable<TCollectionItem> _updatedItems;
 		private readonly UpsertOptions<TCollectionItem, TIdentity> _upsertOptions;
 
-		public UpsertMongoUnitOfWork(TCollectionItem[] updatedCollection, UpsertOptions<TCollectionItem, TIdentity> upsertOptions)
+		public UpsertMongoUnitOfWork(IEnumerable<TCollectionItem> updatedItems, UpsertOptions<TCollectionItem, TIdentity> upsertOptions)
 		{
-			_updatedCollection = updatedCollection;
+			_updatedItems = updatedItems;
 			_upsertOptions = upsertOptions;
 		}
 
-		public override IEnumerable<TCollectionItem> Do(IMongoCollection<TCollectionItem> mongoCollection)
+		public override IEnumerable<TCollectionItem> Do(ILogger logger, IMongoCollection<TCollectionItem> mongoCollection)
 		{
 			var existingItems = mongoCollection.Find(i => true).ToList();
+			var touchedItemIndices = new List<int>();
 
-			var replacements = _updatedCollection.Intersect(existingItems, _upsertOptions.IdentityComparer);
-			foreach (var replacement in replacements)
+			foreach (var updatedItem in _updatedItems)
 			{
-				var existingItem = existingItems.Single(i => _upsertOptions.IdentityComparer.Equals(i, replacement));
+				var existingItem = existingItems.SingleOrDefault(i => _upsertOptions.IdentityComparer.Equals(i, updatedItem));
 
-				var isExistingItemExpired = (_upsertOptions.ActualityComparer?.Compare(existingItem, replacement) ?? -1) < 0;
-
-				if (isExistingItemExpired)
+				if (existingItem == null)
 				{
-					var selectorExpression = _upsertOptions.IdentityComparerExpression(_upsertOptions.IdentitySelector(replacement));
-					replacement.ObjectId = existingItem.ObjectId;
-					mongoCollection.ReplaceOne(selectorExpression, replacement);
+					logger.LogInformation($"{mongoCollection.CollectionNamespace.FullName} MongoDb collection: inserting new item");
+					mongoCollection.InsertOne(updatedItem);
+				}
+				else
+				{
+					touchedItemIndices.Add(existingItems.IndexOf(existingItem));
+
+					var isExistingItemExpired = (_upsertOptions.ActualityComparer?.Compare(existingItem, updatedItem) ?? -1) < 0;
+
+					if (isExistingItemExpired)
+					{
+						var selectorExpression = _upsertOptions.IdentityComparerExpression(_upsertOptions.IdentitySelector(updatedItem));
+						updatedItem.ObjectId = existingItem.ObjectId;
+
+						logger.LogInformation($"{mongoCollection.CollectionNamespace.FullName} MongoDb collection: replacing existing item {existingItem.ObjectId}");
+						mongoCollection.ReplaceOne(selectorExpression, updatedItem);
+					}
+				}
+
+				yield return updatedItem;
+			}
+
+			for (var i = 0; i < existingItems.Count; i++)
+			{
+				if (!touchedItemIndices.Contains(i))
+				{
+					var selectorExpression = _upsertOptions.IdentityComparerExpression(_upsertOptions.IdentitySelector(existingItems[i]));
+
+					logger.LogInformation($"{mongoCollection.CollectionNamespace.FullName} MongoDb collection: deleting item {existingItems[i].ObjectId}");
+					mongoCollection.DeleteOne(selectorExpression);
 				}
 			}
-
-			var redundands = existingItems.Except(_updatedCollection, _upsertOptions.IdentityComparer);
-			foreach (var redundand in redundands)
-			{
-				var selectorExpression = _upsertOptions.IdentityComparerExpression(_upsertOptions.IdentitySelector(redundand));
-				mongoCollection.DeleteOne(selectorExpression);
-			}
-
-			var additions = _updatedCollection.Except(existingItems, _upsertOptions.IdentityComparer).ToList();
-			if (additions.Any())
-			{
-				mongoCollection.InsertMany(additions);
-			}
-
-			return additions.Union(replacements);
 		}
 
 		public override void Dispose()
